fs: sync: Try to insure against data loss with synchronization disabled
Obviously, subject to file synchronization disabled, a huge data loss
will occur after the kernel crash as all the information that was changed
since the last synchronization will be lost. Even considering that file
synchronization will never be used for a long period of time, there is
a risk of such a situation to happen due to different and unexpected
reasons. To make the time interval with sb sync disabled not so dangerous
use a simple module that will handle file synchronization during kernel
panic and reboot events. It can also automatically turn on file sync and
flush all the changed data to disk after LCD panel disablement to make
the risk even less.

If LCD notifier is used, sys_sync() syscall will be used after delay
msecs since the display turns off. This is made to avoid a significant
jitter in case someone wants to turn on the display within few seconds.

Summarizing it all up, this method noticeably reduces the chances of data
loss due to expected system issue. The only weak point here is a hardware
crash.  If a sudden power loss or hardware malfunction occurs, there will
be nothing, but emptiness on the storage.

As there is no point in having an accessible file synchronization trigger
anymore, make it read-only.  The only function that module param can
accomplish now is to show the current state of superblock synchronization.
Dynamic way provides the same functionality with lower risks anyway.

Signed-off-by: Alex Saiko <solcmdr@gmail.com>